Use floats instead of doubles where it makes sense.
git-svn-id: https://mumble.svn.sourceforge.net/svnroot/mumble/trunk@966 05730e5d-ab1b-0410-a4ac-84af385074fa

+++ src/mumble/ASIOInput.cpp
@@ -538,7 +538,7 @@ ASIOInput::bufferReady(long buffindex) {
 	for (c=0;c<iNumMic;c++)
 		addBuffer(aciInfo[c].type, abiInfo[c].buffers[buffindex], pdInputBuffer);
 
-	double mul = 1.0 / (iNumMic);
+	float mul = 1.0 / (iNumMic);
 
 	spx_uint32_t inlen, outlen;
