+++ src/mumble/smallft.cpp
@@ -38,9 +38,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
  ********************************************************************/
 
-#include <stdlib.h>
-#include <string.h>
-#include <math.h>
+#include "mumble_pch.hpp"
 #include "smallft.h"
 
 static void drfti1(int n, float *wa, int *ifac){
@@ -104,8 +102,8 @@ static void drfti1(int n, float *wa, int *ifac){
       for (ii=2;ii<ido;ii+=2){
 	fi+=1.f;
 	arg=fi*argld;
-	wa[i++]=cos(arg);
-	wa[i++]=sin(arg);
+	wa[i++]=cosf(arg);
+	wa[i++]=sinf(arg);
       }
       is+=ido;
     }
@@ -288,8 +286,8 @@ static void dradfg(int ido,int ip,int l1,int idl1,float *cc,float *c1,
   int idp2,ipp2;
 
   arg=tpi/(ip * 1.f);
-  dcp=cos(arg);
-  dsp=sin(arg);
+  dcp=cosf(arg);
+  dsp=sinf(arg);
   ipph=(ip+1)>>1;
   ipp2=ip;
   idp2=ido;
@@ -858,8 +856,8 @@ static void dradbg(int ido,int ip,int l1,int idl1,float *cc,float *c1,
   t10=ip*ido;
   t0=l1*ido;
   arg=tpi/(ip * 1.f);
-  dcp=cos(arg);
-  dsp=sin(arg);
+  dcp=cosf(arg);
+  dsp=sinf(arg);
   nbd=(ido-1)>>1;
   ipp2=ip;
   ipph=(ip+1)>>1;
